+++ app/assets/scripts/selfie/campaign.js
@@ -9,7 +9,10 @@ function( angular , c6State  , PaginatedListState                    ,
         equals = angular.equals,
         extend = angular.extend,
         forEach = angular.forEach,
-        isObject = angular.isObject;
+        isObject = angular.isObject,
+        isArray = angular.isArray,
+        isFunction = angular.isFunction,
+        isDate = angular.isDate;
 
     function deepExtend(target, extension) {
         forEach(extension, function(extensionValue, prop) {
@@ -23,6 +26,41 @@ function( angular , c6State  , PaginatedListState                    ,
         });
 
         return target;
+    }
+
+    /* Adapted from Angular 1.4.7, modified to not handle regular expressions.
+        Used for merging objects. */
+    function baseExtend(dst, objs, deep) {
+        for (var i = 0, ii = objs.length; i < ii; ++i) {
+            var obj = objs[i];
+            if (!isObject(obj) && !isFunction(obj)) {
+                continue;
+            }
+            var keys = Object.keys(obj);
+            for (var j = 0, jj = keys.length; j < jj; j++) {
+                var key = keys[j];
+                var src = obj[key];
+                if (deep && isObject(src)) {
+                    if (isDate(src)) {
+                        dst[key] = new Date(src.valueOf());
+                    } else {
+                        if (!isObject(dst[key])) {
+                            dst[key] = isArray(src) ? [] : {};
+                        }
+                        baseExtend(dst[key], [src], true);
+                    }
+                } else {
+                    dst[key] = src;
+                }
+            }
+        }
+        return dst;
+    }
+
+    /* Adapted from Angular 1.4.7 as the currently used version of Angular
+        does not include a merge function. */
+    function merge(dst) {
+        return baseExtend(dst, [].slice.call(arguments, 1), true);
     }
 
     return angular.module('c6.app.selfie.campaign', [c6State.name])
@@ -75,8 +113,10 @@ function( angular , c6State  , PaginatedListState                    ,
 
         .controller('SelfieCampaignsController', ['$injector','$scope','$q','cState',
                                                   'ConfirmDialogService','ThumbnailService',
+                                                  'CampaignService',
         function                                 ( $injector , $scope , $q , cState ,
-                                                   ConfirmDialogService , ThumbnailService ) {
+                                                   ConfirmDialogService , ThumbnailService ,
+                                                   CampaignService ) {
             var SelfieCampaignsCtrl = this;
 
             $injector.invoke(PaginatedListController, this, {
@@ -102,8 +142,54 @@ function( angular , c6State  , PaginatedListState                    ,
                 return $q.when(null);
             }
 
+            function addMetaData() {
+                var Ctrl = SelfieCampaignsCtrl,
+                    model = Ctrl.model.items.value,
+                    ids = model.map(function(campaign) {
+                        return campaign.id;
+                    }).join(',');
+
+                Ctrl.metaData = model.reduce(function(result, campaign) {
+                    var card = campaign.cards && campaign.cards[0];
+
+                    if (!card) { return result; }
+
+                    result[campaign.id] = {
+                        sponsor: card.params.sponsor,
+                        logo: card.collateral.logo
+                    };
+
+                    thumbFor(card).then(function(thumb) {
+                        result[campaign.id].thumb = thumb;
+                    });
+
+                    return result;
+                },{});
+
+                if (ids) {
+                    CampaignService.getAnalytics(ids)
+                        .then(function(stats) {
+                            stats.forEach(function(stat) {
+                                var campaignId = stat.campaignId;
+
+                                if (!campaignId || !Ctrl.metaData[campaignId]) {
+                                    return;
+                                }
+
+                                Ctrl.metaData[campaignId].stats = {
+                                    views: stat.summary.views,
+                                    spend: stat.summary.totalSpend
+                                };
+                            });
+                        });
+                }
+            }
+
             this.initWithModel = function(model) {
                 this.model = model;
+
+                addMetaData();
+                model.on('PaginatedListHasUpdated', addMetaData);
 
                 this.filters = [
                     'draft',
@@ -162,28 +248,6 @@ function( angular , c6State  , PaginatedListState                    ,
                     return filter.checked ? filters.concat(filter.id) : filters;
                 },[]).join(',');
             };
-
-            $scope.$watch(function() {
-                return SelfieCampaignsCtrl.model.items.value;
-            }, function(model) {
-
-                SelfieCampaignsCtrl.metaData = model.reduce(function(result, campaign) {
-                    var card = campaign.cards && campaign.cards[0];
-
-                    if (!card) { return result; }
-
-                    result[campaign.id] = {
-                        sponsor: card.params.sponsor,
-                        logo: card.collateral.logo
-                    };
-
-                    thumbFor(card).then(function(thumb) {
-                        result[campaign.id].thumb = thumb;
-                    });
-
-                    return result;
-                },{});
-            });
         }])
 
         .config(['c6StateProvider',
@@ -231,9 +295,9 @@ function( angular , c6State  , PaginatedListState                    ,
         .config(['c6StateProvider',
         function( c6StateProvider ) {
             c6StateProvider.state('Selfie:Campaign', ['cinema6','SelfieLogoService',
-                                                      'c6State','$q',
+                                                      'c6State','$q','ConfirmDialogService',
             function                                 ( cinema6 , SelfieLogoService ,
-                                                       c6State , $q ) {
+                                                       c6State , $q , ConfirmDialogService ) {
                 var SelfieState = c6State.get('Selfie');
 
                 this.templateUrl = 'views/selfie/campaigns/campaign.html';
@@ -271,6 +335,37 @@ function( angular , c6State  , PaginatedListState                    ,
 
                     this.campaign.paymentMethod = this.campaign.paymentMethod ||
                         primaryPaymentMethod.token;
+                };
+
+                this.exit = function() {
+                    if (this._campaign.status === 'draft') {
+                        return this.saveCampaign()
+                            .catch(function() {
+                                var deferred = $q.defer();
+
+                                ConfirmDialogService.display({
+                                    prompt: 'There was a problem saving your campaign, would ' +
+                                        'you like to stay on this page to edit the campaign?',
+                                    affirm: 'Yes, stay on this page',
+                                    cancel: 'No',
+
+                                    onCancel: function() {
+                                        deferred.resolve();
+
+                                        return ConfirmDialogService.close();
+                                    },
+                                    onAffirm: function() {
+                                        deferred.reject();
+
+                                        return ConfirmDialogService.close();
+                                    }
+                                });
+
+                                return deferred.promise;
+                            });
+                    } else {
+                        return $q.when(null);
+                    }
                 };
 
                 this.saveCampaign = function() {
@@ -992,10 +1087,10 @@ function( angular , c6State  , PaginatedListState                    ,
             }]);
         }])
 
-        .controller('SelfieManageCampaignController', ['$scope','cState','c6State',
-                                                       'c6AsyncQueue','CampaignService',
-        function                                      ( $scope , cState , c6State ,
-                                                        c6AsyncQueue , CampaignService ) {
+        .controller('SelfieManageCampaignController', ['$scope','cState','c6AsyncQueue',
+                                                       'c6State', 'CampaignService',
+        function                                      ( $scope , cState , c6AsyncQueue ,
+                                                        c6State ,  CampaignService ) {
             var queue = c6AsyncQueue();
 
             Object.defineProperties(this, {
@@ -1017,10 +1112,12 @@ function( angular , c6State  , PaginatedListState                    ,
             // };
 
             this.initWithModel = function(model) {
+                var user = c6State.get('Selfie').cModel;
                 this.card = cState.card;
                 this.campaign = cState.campaign;
                 this.categories = model.categories;
                 this.paymentMethods = model.paymentMethods;
+                this.showAdminTab = (user.entitlements.adminCampaigns === true);
             };
 
             this.update = queue.debounce(function() {
@@ -1065,5 +1162,88 @@ function( angular , c6State  , PaginatedListState                    ,
             c6StateProvider.state('Selfie:Manage:Campaign:Payment', [function() {
                 this.templateUrl = 'views/selfie/campaigns/manage/payment.html';
             }]);
+        }])
+
+        .config(['c6StateProvider',
+        function( c6StateProvider ) {
+            c6StateProvider.state('Selfie:Manage:Campaign:Admin', ['cinema6', '$q',
+            function                                              ( cinema6 ,  $q ) {
+                this.templateUrl = 'views/selfie/campaigns/manage/admin.html';
+                this.controller = 'SelfieManageCampaignAdminController';
+                this.controllerAs = 'SelfieManageCampaignAdminCtrl';
+
+                this.campaign = null;
+
+                this.beforeModel = function() {
+                    this.campaign = this.cParent.campaign;
+                };
+
+                this.model = function() {
+                    var model = {
+                        updateRequest: null
+                    };
+                    if(this.campaign.updateRequest) {
+                        var updateHash = this.campaign.id + ':' + this.campaign.updateRequest;
+                        model.updateRequest = cinema6.db.find('updateRequest', updateHash);
+                    }
+                    return $q.all(model);
+                };
+            }]);
+        }])
+
+        .controller('SelfieManageCampaignAdminController', ['c6State', 'cState', 'cinema6',
+                                                            '$scope', 'c6Debounce',
+        function                                           ( c6State ,  cState ,  cinema6 ,
+                                                             $scope ,  c6Debounce ){
+            var self = this;
+            var updateRequest;
+
+            this.initWithModel = function(model) {
+                updateRequest = model.updateRequest;
+                extend(self, {
+                    showApproval: false,
+                    campaign: cState.campaign.pojoify(),
+                    updatedCampaign: cState.campaign.pojoify(),
+                    previewCard: null,
+                    rejectionReason: ''
+                });
+                if(updateRequest) {
+                    var updates = updateRequest.data;
+                    merge(self.updatedCampaign, updates);
+                    extend(self, {
+                        showApproval: true,
+                        previewCard: copy(self.updatedCampaign.cards[0])
+                    });
+                }
+            };
+
+            this.approveCampaign = function() {
+                updateRequest.data = self.updatedCampaign;
+                updateRequest.status = 'approved';
+                updateRequest.save().then(function() {
+                    c6State.goTo('Selfie:CampaignDashboard');
+                });
+            };
+
+            this.rejectCampaign = function() {
+                extend(updateRequest, {
+                    status: 'rejected',
+                    rejectionReason: self.rejectionReason
+                }).save().then(function() {
+                    c6State.goTo('Selfie:CampaignDashboard');
+                });
+            };
+
+            this._loadPreview = c6Debounce(function(args) {
+                var _card = copy(args[0]);
+                _card.params.sponsor = self.updatedCampaign.advertiserDisplayName;
+                self.previewCard = _card;
+            }, 2000);
+
+            $scope.$watch(function() {
+                return self.updatedCampaign.cards[0];
+            }, function(card) {
+                self._loadPreview(card);
+            }, true);
         }]);
 });