+++ test/spec/c6_sref.directive.ut.js
@@ -137,12 +137,6 @@
                     });
                 });
 
-                it('should not call c6State.goTo() when clicked', function() {
-                    $sref.click();
-
-                    expect(c6State.goTo).not.toHaveBeenCalled();
-                });
-
                 describe('if the current state has the same URL as the one being referenced', function() {
                     beforeEach(function() {
                         Object.defineProperty(c6State, 'current', {